+++ tests/Feature/TaskControllerTest.php
@@ -8,16 +8,25 @@ use Database\Seeders\{StatusesTableSeeder, TasksTableSeeder};
 
 class TaskControllerTest extends TestCase
 {
-    public User $user;
-    public Task $task;
+    /**
+     *
+     * @var ?User
+     */
+    public $user;
+
+    /**
+     *
+     * @var ?Task
+     */
+    public $task;
 
     public function setUp(): void
     {
         parent::setUp();
         $this->seed(StatusesTableSeeder::class);
         $this->seed(TasksTableSeeder::class);
-        $this->user = User::find(2);
-        $this->task = Task::first();
+        $this->user = User::findOrFail(2);
+        $this->task = Task::firstOrFail();
     }
 
     public function testIndex(): void
@@ -28,12 +37,20 @@ class TaskControllerTest extends TestCase
 
     public function testShow(): void
     {
+        self::assertTrue(isset($this->task));
+        if (is_null($this->task)) {
+            return;
+        }
         $response = $this->get(route('tasks.show', $this->task));
         $response->assertOk();
     }
 
     public function testCreate(): void
     {
+        self::assertTrue(isset($this->user));
+        if (is_null($this->user)) {
+            return;
+        }
         $response = $this->get(route('tasks.create'));
         $response->assertForbidden();
         // with authentication
@@ -43,6 +60,10 @@ class TaskControllerTest extends TestCase
 
     public function testEdit(): void
     {
+        self::assertTrue(isset($this->user));
+        if (is_null($this->user)) {
+            return;
+        }
         $response = $this->get(route('tasks.edit', 1));
         $response->assertForbidden();
         // with authentication
@@ -52,6 +73,10 @@ class TaskControllerTest extends TestCase
 
     public function testStore(): void
     {
+        self::assertTrue(isset($this->user));
+        if (is_null($this->user)) {
+            return;
+        }
         $data = ["name" => "testing", 'created_by_id' => 2, 'assigned_to_id' => 1, 'status_id' => 1];
         $response = $this->post(route('tasks.store'), $data);
         $response->assertSessionHasNoErrors();
@@ -66,6 +91,11 @@ class TaskControllerTest extends TestCase
 
     public function testUpdate(): void
     {
+        self::assertTrue(isset($this->user));
+        self::assertTrue(isset($this->task));
+        if (is_null($this->user) || is_null($this->task)) {
+            return;
+        }
         $this->task->name = 'testing';
         $data = $this->task->only("name", "description", "status_id", "assigned_to_id", "created_by_id");
         $response = $this->patch(route('tasks.update', $this->task), $data);
@@ -81,6 +111,11 @@ class TaskControllerTest extends TestCase
 
     public function testDestroy(): void
     {
+        self::assertTrue(isset($this->user));
+        self::assertTrue(isset($this->task));
+        if (is_null($this->user) || is_null($this->task)) {
+            return;
+        }
         $data = ['id' => $this->task->id];
         $response = $this->delete(route('tasks.destroy', $this->task->id));
         $response->assertSessionHasNoErrors();
